Updated migration script
Migration script now runs start to finish. However, I'm now convinced that reduceVar(do_jk = TRUE) is a much better way to handle parsimony for these variables, and will be easier to explain (as it actually prioritizes model performance, unlike VSURF). Committing before I attempt to implement this for the migratory model. Note that I will have to write this out manually for the residential model, but I don't think this should be too hard.