First shot at a new draft
Methods are about done, as well as the appendix and the outer framework of the results. Next I need to start rerunning the figures and transcribing the results into the new paper. Note that I have also promised to add variable importance scores, which should be pretty straightforward but would be a new figure for the paper (will go in place of Fig 2, which is moving to the appendix).